MNIST/requirements.txt
matplotlib
numpy
pandas
scikit-learn
scipy
tensorflow-gpu
